Annotate m/z values with glycan compositions based on theoretical values

glycoAnnotate() annotates peaks or features in MS data, using either a pre-generated table by glycoPredict or by generating a new table.

Usage

glycoAnnotate(
  data,
  mz_column = "mz",
  mzmin_column = NULL,
  mzmax_column = NULL,
  pred_table = NULL,
  param = NULL,
  collapse = F,
  collapse_columns = NULL,
  error = 3,
  error_units = "ppm"
)

Slots

data

Dataframe containing data to be annotated. For example, feature dataframe from XCMS pre-processing (LC-MS or direct inject) or features from Cardinal (MALDI).

mz_column

Name of column containing m/z values.

mzmin_column

OPTIONAL: Name of column containing minimum m/z data values. If supplied, will do overlap-overlap matching. Generally only if mzmin and mzmax values generated during peak picking. If not provided, mz value will be annotated if within range of theoretical mz +- error.

mzmax_column

OPTIONAL: Name of column containing maximum m/z data values. If supplied, will do overlap-overlap matching. Generally only if mzmin and mzmax values generated during peak picking.If not provided, mz value will be annotated if within range of theoretical mz +- error.

pred_table

Table generated previously by glycoPredict. MUST BE LONG FORMAT - select format='long' when running prediction.

param

glycoPredictParam object for generation of table of theoretical mz values for annotation.

collapse

Logical. If TRUE, annotations will be 'collapsed' so that multiple annotations for one mz will be in the same row, comma separated (nrow of output is in this case equal to nrow of input data). If FALSE (default), it is possible that rows in the input dataframe are repeated with different annotations. The information on annotations is more detailed in this case. Collapsing can also be done afterwards on the output using glycoAnnotationsCollapse.

collapse_columns

Columns to be pasted together before collapsing. Only needed if collapse=TRUE and non-default columns wanted - default is molecule name and ion. If prediction table provided to pred_table instead of param, column names are required.

error

Numeric value - error used to create window for matching. mz values will be matched against theoretical mzs +- error.

error_units

Units for error - can be 'ppm' or 'Da'

See also

Examples

#with prediction parameters
gpp <- glycoPredictParam(dp = c(1, 8), modifications = "deoxy", polarity = "pos", naming = "IUPAC")
annotated_data <- glycoAnnotate(data = data, param = gpp, error = 1.5, units = 'ppm', collapse = T)
#> Error in glycoAnnotate(data = data, param = gpp, error = 1.5, units = "ppm",     collapse = T): unused argument (units = "ppm")

#with prediction table
gpp <- glycoPredictParam(dp = c(1, 8), modifications = "deoxy", polarity = "pos",  naming = "IUPAC")
pred_table <- glycoPredict(param = gpp)
#> Formal class 'glycoPredictParam' [package "GlycoAnnotateR"] with 14 slots
#>   ..@ dp                 : num [1:2] 1 8
#>   ..@ polarity           : chr "pos"
#>   ..@ scan_range         : num [1:2] 175 1400
#>   ..@ pent_option        : logi FALSE
#>   ..@ modifications      : chr "deoxy"
#>   ..@ nmod_max           : num 1
#>   ..@ double_sulphate    : logi FALSE
#>   ..@ label              : chr "none"
#>   ..@ ion_type           : chr "ESI"
#>   ..@ format             : chr "long"
#>   ..@ adducts            : chr "all"
#>   ..@ naming             : chr "IUPAC"
#>   ..@ glycan_linkage     : chr "none"
#>   ..@ modification_limits: chr "none"
#> Glycans will be predicted according to the following glycoPredictParam() object:
#>  
annotated_data <- glycoAnnotate(data = data, pred_table = pred_table, error = 1.5, units = 'ppm', collapse = T, collapse_columns = c("IUPAC name", "ion"))
#> Error in glycoAnnotate(data = data, pred_table = pred_table, error = 1.5,     units = "ppm", collapse = T, collapse_columns = c("IUPAC name",         "ion")): unused argument (units = "ppm")