Collapse annotated m/z values to have one row per peak/feature

glycoAnnotationsCollapse() collapses the output of glycoAnnotate in the case of multiple annotations per peak or feature so that there is one row per peak/feature with multiple annotations comma-separated.

Usage

glycoAnnotationsCollapse(annotated_data, collapse_columns, noncollapse_columns)

Slots

annotated_data

Dataframe annotated by glycoAnnotate that has NOT been collapsed and has multiple annotations per peak/feature.

collapse_columns

Names of columns to be pasted together before collapsing. Suggested is molecule name and ion.

noncollapse_columns

Names of columns that uniquely identify peaks and that should be retained after collapsing - these are generally the column names of your input dataframe before annotation.

See also

Examples

#annotate dataframe
gpp <- glycoPredictParam(dp = c(1, 8), modifications = "deoxy", polarity = "pos", naming = "IUPAC")
annotated_data <- glycoAnnotate(data = data, param = gpp, error = 1.5, units = 'ppm', collapse = F)
#> Error in glycoAnnotate(data = data, param = gpp, error = 1.5, units = "ppm",     collapse = F): unused argument (units = "ppm")

#collapse multiple annotations
annotated_data_collapsed <- glycoAnnotationsCollapse(annotated_data = annotated_data, collapse_columns = c('IUPAC name', 'ion'), noncollapse_columns = c('mz', 'rt', 'sampleA', 'sampleB'))
#> Error in eval(expr, envir, enclos): object 'annotated_data' not found