glycoPredictParam class

An S4 class that contains the parameters for glycoPredict() function.

Details

glycoPredict() is used to predict masses and mass to charge ratios of all theoretically possible glycans within a set of constraining parameters (defined in the glycoPredictParam object). This package was written for annotation of mass spec data (especially LC-MS) but if used for other purposes either ionisation mode and very wide scan ranges can be given. The function works by sourcing a python file and then using the function encoded in the python script.

Slots

dp

Degree of polymerisation range. It must be a numeric vector of length 2. Default: c(1,6).

polarity

Ionisation mode used. Accepts 'pos' or 'neg' (default).

scan_range

Scan range used during MS. It must be a numeric vector of length 2. Default: c(175, 1400).

pent_option

Logical. Should pentose monomers be included? Default: FALSE.

modifications

Modifications to be considered. Any combination of 'carboxylicacid', 'sialicacid', 'phosphate', 'deoxy', 'nacetyl', 'omethyl', 'anhydrobridge', 'oacetyl', 'unsaturated', 'alditol', 'amino', 'dehydrated', 'sulphate', 'aminopentyllinker' or 'all' or 'none' (default)

nmod_max

Maximum number of modifications per monomer, calculated by the number of modifications over the number of monomers (default 1). Does not take into account unsaturated, alditol or dehydrated.

double_sulphate

Logical. Can monomers be double-sulphated. If TRUE, nmod_max needs to have a value of at least 2.

label

Are sugars labelled by reductive amination? Current supported labels are: "none", "procainamide","2-aminobenzoic acid", "2-aminobenzamide", "1-phenyl-3-methyl-5-pyrazolone".

ion_type

Ionisation type. Currently accepted ESI and MALDI. Impacts ions.

naming

Notation for molecule names. Uses commonly accepted abbreviations. Possibilities: 'IUPAC' (default), 'Oxford', 'GlycoCT'

glycan_linkage

Option to implement filters for O- and N-glycans. Possibilities: 'none' (default), 'nglycan' or 'oglycan'.

modification_limits

Option to implement user created filters. Must be a named list, with names as modifications and values as limits.

See also

Examples

gpp <- glycoPredictParam()
gpp@dp <- c(1,7)
gpp@polarity <- 'neg'
gpp@scan_range <- c(150, 1300)
gpp@modifications <- c('sulphate', 'carboxylicacid')
gpp@double_sulphate <- TRUE
predicted.df <- glycoPredict(param = gpp)
#> Formal class 'glycoPredictParam' [package "GlycoAnnotateR"] with 14 slots
#>   ..@ dp                 : num [1:2] 1 7
#>   ..@ polarity           : chr "neg"
#>   ..@ scan_range         : num [1:2] 150 1300
#>   ..@ pent_option        : logi FALSE
#>   ..@ modifications      : chr [1:2] "sulphate" "carboxylicacid"
#>   ..@ nmod_max           : num 1
#>   ..@ double_sulphate    : logi TRUE
#>   ..@ label              : chr "none"
#>   ..@ ion_type           : chr "ESI"
#>   ..@ format             : chr "long"
#>   ..@ adducts            : chr "all"
#>   ..@ naming             : chr "IUPAC"
#>   ..@ glycan_linkage     : chr "none"
#>   ..@ modification_limits: chr "none"
#> Glycans will be predicted according to the following glycoPredictParam() object:
#>